requirements.txt
asgiref==3.7.2
Django==4.2.3
django-admin-honeypot==1.1.0
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.7
importlib-metadata==6.8.0
inflection==0.5.1
Markdown==3.4.4
packaging==23.1
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
zipp==3.16.2
